cricketinfo/requirements.txt
Django==2.1.5
djangorestframework==3.9.4
gunicorn
pillow
requests